Move fetch logic into separate module
